Removed all argparse. Changed package capitalisation
