+++ HCPInterface/hcp/errors.py
